test: split test into their own test files
